feat(web): 상품상세 페이지 레이아웃 컴포넌트 분리 및 댓글섹션 컴포넌트 파일 분리
