+++ apps/web/src/app/(mobile)/products/[id]/comment/components/CommentLayout.tsx
@@ -115,8 +115,8 @@ const CommentListSkeleton = () => {
       <div className="flex flex-col items-center gap-y-3">
         <BubbleChatFill />
         <div className="flex flex-col items-center gap-y-1">
-          <p className="font-semibold text-gray-700">가장 먼저 댓글을 달아보세요!</p>
-          <p className="text-sm font-medium text-gray-500">핫딜을 주제로 소통해요</p>
+          <p className="font-semibold text-gray-700">첫 후기를 남겨주세요!</p>
+          <p className="text-sm font-medium text-gray-500">댓글로 함께 소통해요</p>
         </div>
       </div>
     </div>
