+++ apps/web/src/app/(mobile)/products/[id]/components/ProductDetailContainerServer.tsx
@@ -1,9 +1,7 @@
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
-import BasicLayout from '@/components/layout/BasicLayout';
 import { ProductQueries } from '@/entities/product';
 
-import ProductDetailPageHeader from './ProductDeatilPageHeader';
 import ProductDetailContainer from './ProductDetailContainer';
 
 const ProductDetailContainerServer = async ({ productId }: { productId: number }) => {
@@ -11,9 +9,7 @@ const ProductDetailContainerServer = async ({ productId }: { productId: number }
   await queryClient.prefetchQuery(ProductQueries.productServer({ id: productId }));
   return (
     <HydrationBoundary state={dehydrate(queryClient)}>
-      <BasicLayout header={<ProductDetailPageHeader productId={productId} />}>
-        <ProductDetailContainer productId={productId} />
-      </BasicLayout>
+      <ProductDetailContainer productId={productId} />
     </HydrationBoundary>
   );
 };
