+++ apps/web/src/app/(mobile)/products/[id]/components/comment/index.tsx
@@ -1,0 +1,3 @@
+export { default as CommentList } from './CommentList';
+export { default as CommentListSkeleton } from './CommentListSkeleton';
+export { default as CommentSection } from './CommentSection';
